Initial commit
This project implements monte carlo for approximation of pi. This is done by a ratio of random points inside a circle circumscribed by a square to the sum of random points outside.

This is currently executed as 1 evaluation per thread per block on the GPU. There is much room for optimization.